transactions

+++ src/main/java/com/developer/library/model/Users.java
@@ -10,7 +10,8 @@ import org.springframework.stereotype.Component;
 public class Users {
 
     @Id
-    private int id;
+    private long id;
+
     private String name;
 
     @Column(unique = true)
@@ -25,11 +26,11 @@ public class Users {
         this.email = email;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
